+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### December
 
+- Add an AppImage manager
+- Add a LaTeX Editor
 - GNOME Updates
   - Change terminal hotkeys
   - Use [`copyous`](https://github.com/boerdereinar/copyous) instead of `pano` for clipboard